Unlit.js -> html-utils.js, move htmlToElement to html-utils
