+++ src/fontra/client/web-components/icon-button.js
@@ -1,6 +1,6 @@
 import { InlineSVG } from "./inline-svg.js";
-import * as html from "/core/unlit.js";
-import { UnlitElement } from "/core/unlit.js";
+import * as html from "/core/html-utils.js";
+import { UnlitElement } from "/core/html-utils.js";
 
 export class IconButton extends UnlitElement {
   static styles = `
