+++ src/fontra/views/editor/panel-designspace-navigation.js
@@ -1,12 +1,12 @@
 import { css } from "../third-party/lit.js";
 import { getAxisBaseName } from "/core/glyph-controller.js";
+import * as html from "/core/html-utils.js";
+import { htmlToElement } from "/core/html-utils.js";
 import { controllerKey, ObservableController } from "/core/observable-object.js";
-import * as html from "/core/unlit.js";
 import {
   boolInt,
   enumerate,
   escapeHTMLCharacters,
-  htmlToElement,
   objectsEqual,
   range,
   rgbaToCSS,
