+++ src/fontra/views/editor/panel-glyph-search.js
@@ -1,6 +1,6 @@
 import { css } from "../third-party/lit.js";
 import Panel from "./panel.js";
-import * as html from "/core/unlit.js";
+import * as html from "/core/html-utils.js";
 
 export default class GlyphSearchPanel extends Panel {
   identifier = "glyph-search";
