+++ src/fontra/views/editor/panel-reference-font.js
@@ -1,9 +1,9 @@
 import { css } from "../third-party/lit.js";
 import Panel from "./panel.js";
-import * as html from "/core/unlit.js";
+import * as html from "/core/html-utils.js";
 
+import { div, input, label, option, select } from "/core/html-utils.js";
 import { ObservableController } from "/core/observable-object.js";
-import { div, input, label, option, select } from "/core/unlit.js";
 import { fetchJSON, fileNameExtension, withTimeout } from "/core/utils.js";
 import { dialog } from "/web-components/modal-dialog.js";
 import { UIList } from "/web-components/ui-list.js";
