+++ src/fontra/views/editor/sidebar.js
@@ -1,5 +1,5 @@
 import { clamp } from "../../core/utils.js";
-import * as html from "/core/unlit.js";
+import * as html from "/core/html-utils.js";
 
 export const MIN_SIDEBAR_WIDTH = 200;
 export const MAX_SIDEBAR_WIDTH = 500;
